chore(app): migrates routes to pages
