+++ app/src/pages/login.tsx
@@ -1,3 +1,4 @@
+import { Label } from "@radix-ui/react-label";
 import { Button } from "@shadcn-ui/components/ui/button";
 import {
   Card,
@@ -7,28 +8,23 @@ import {
   CardTitle,
 } from "@shadcn-ui/components/ui/card";
 import { Input } from "@shadcn-ui/components/ui/input";
-import { Label } from "@shadcn-ui/components/ui/label";
-import { createFileRoute, Link } from "@tanstack/react-router";
 import { useRef } from "react";
+import { Link } from "react-router";
 import { useAuth } from "../providers/auth/useAuth";
 
-export const Route = createFileRoute("/signup")({
-  component: RouteComponent,
-});
+export const LoginPage = () => {
+  const auth = useAuth();
 
-function RouteComponent() {
   const emailInput = useRef<HTMLInputElement>(null);
   const passwordInput = useRef<HTMLInputElement>(null);
 
-  const auth = useAuth();
-
-  const signup = () => {
+  const login = () => {
     const email = emailInput.current?.value;
     const password = passwordInput.current?.value;
 
     if (!email || !password) return;
 
-    auth.signup({ email, password });
+    auth.login({ email, password });
   };
 
   return (
@@ -37,8 +33,10 @@ function RouteComponent() {
         <div className="flex flex-col gap-6">
           <Card>
             <CardHeader>
-              <CardTitle>Create your user</CardTitle>
-              <CardDescription>Sign up and create a profile</CardDescription>
+              <CardTitle>Login to your account</CardTitle>
+              <CardDescription>
+                Enter your email below to login to your account
+              </CardDescription>
             </CardHeader>
             <CardContent>
               <div className="flex flex-col gap-6">
@@ -62,15 +60,15 @@ function RouteComponent() {
                   />
                 </div>
                 <div className="flex flex-col gap-3">
-                  <Button onClick={() => signup()} className="w-full">
-                    Signup
+                  <Button onClick={() => login()} className="w-full">
+                    Login
                   </Button>
                 </div>
               </div>
               <div className="mt-4 text-center text-sm">
                 Already have an account?{" "}
-                <Link to="/login" className="underline underline-offset-4">
-                  Login
+                <Link to="/signup" className="underline underline-offset-4">
+                  Signup
                 </Link>
               </div>
             </CardContent>
@@ -79,4 +77,4 @@ function RouteComponent() {
       </div>
     </div>
   );
-}
+};
